// web/src/app/resources/3-public/member-verify/member-verify.types.ts
export interface VerifiedMemberData {
    id: string | number;
    code: string;
    name_kh: string;
    name_en: string;
    gender: string;
    nationality: string;
    dob: string;
    pob: string;
    phone: string;
    email: string;
    organization_kh: string;
    organization_en: string;
    role_title: string;
    status: string;
    valid_until: string;
    avatar_url: string;
}